+++ crates/hipparchus-mean/src/sequence.rs
@@ -1,23 +1,33 @@
-use std::{ops::{Add, Sub, Mul, Div}, iter::repeat_with};
+use std::ops::{Add, Sub, Mul, Div};
+use std::iter::repeat_with;
 use num::{FromPrimitive, Zero, One};
 
 #[repr(i32)]
 pub enum Sequence<T> where
     T: Add + Sub + Mul + Div<Output=T> + FromPrimitive + Zero + One + Copy + Clone,
 {
+    // 
     Arithmetic { init:T, difference:T } = 1,
     Natural(bool) = 2,
     Odd = 3,
     Even(bool) = 4,
     Geometric { init:T, ratio:T } = 5,
-    Fibonacci = 6,
-    Lucas = 7,
-    Padova = 8,
-    Catalan = 9,
-    LookAndSay(usize) = 10,
+
+    // 
+    Triangular = 10,
+    Square = 11,
+    Cubic = 12,
+    Harmonic { init:T, difference:T } = 13,
+
+    // 
+    Fibonacci = 20,
+    Lucas = 21,
+    Padova = 22,
+    Catalan = 29,
+    LookAndSay(usize) = 30,
 }
 
-impl<'a, T> Sequence<T> where
+impl<T> Sequence<T> where
     T: Add + Sub + Mul + Div<Output=T> + FromPrimitive + Zero + One + Copy + Clone,
 {
     pub fn vec(self, n:usize) -> Vec<T>
@@ -41,6 +51,26 @@ impl<'a, T> Sequence<T> where
                     T::from_i32(2).unwrap()
                 )).take(n).collect(),
             Sequence::Geometric { init, ratio } => repeat_with(Self::geometric(init, ratio)).take(n).collect(),
+            Sequence::Triangular => Sequence::Natural(false).map
+            (
+                n, |x|
+                (T::one() + x) * x / T::from_i32(2).unwrap()
+            ),
+            Sequence::Square => Sequence::Natural(false).map
+            (
+                n, |x|
+                x * x
+            ),
+            Sequence::Cubic => Sequence::Natural(false).map
+            (
+                n, |x|
+                x * x * x
+            ),
+            Sequence::Harmonic { init, difference } => Sequence::Arithmetic { init, difference }.map
+            (
+                n, |x|
+                T::one() / x
+            ),
             Sequence::Fibonacci => repeat_with(Self::fibonacci_lucas
                 (
                     T::zero(),
@@ -55,6 +85,42 @@ impl<'a, T> Sequence<T> where
             Sequence::Catalan => repeat_with(Self::catalan()).take(n).collect(),
             Sequence::LookAndSay ( with ) => repeat_with(Self::lookandsay(with)).take(n).collect(),
         }
+    }
+
+    pub fn map<F>(self, n:usize, f:F) -> Vec<T>
+        where F: Fn(T) -> T
+    {
+        let me = self.vec(n);
+        let mut index = 0;
+
+        repeat_with(move ||
+        {
+            let mapped = f(me[index]);
+            index += 1;
+            mapped
+        }).take(n).collect()
+    }
+
+    pub fn fold<F>(self, n:usize, start:i32, end:i32, f:F) -> Vec<T>
+        where F: Fn(&[Option<T>]) -> T
+    {
+        let me = self.vec(n);
+        let mut index = 0;
+
+        repeat_with(move ||
+        {
+            let mut data = Vec::<Option<T>>::new();
+            let a = index + start;
+            let b = index + end;
+            for k in a..b
+            {
+                data.push( if k < 0 || k >= n as i32 { None } else { Some(me[k as usize]) } )
+            }
+
+            let mapped = f(&data);
+            index += 1;
+            mapped
+        }).take(n).collect()
     }
 
     fn arithmetic(init:T, difference:T) -> impl FnMut() -> T
@@ -232,6 +298,39 @@ mod tests
     }
 
     #[test]
+    fn test_sequence_triangular_i32()
+    {
+        let expected = vec!
+        [
+            1,      3,      6,      10,     15, 
+            21,     28,     36,     45,     55, 
+            66,     78,     91,     105,    120, 
+            136,    153,    171,    190,    210, 
+        ];
+        let n = expected.len();
+        let actual = Sequence::Triangular.vec(n);
+        assert_eq!(expected, actual);
+    }
+
+    #[test]
+    fn test_sequence_square_i32()
+    {
+        let expected = vec![1, 4, 9, 16, 25];
+        let n = expected.len();
+        let actual = Sequence::Square.vec(n);
+        assert_eq!(expected, actual);
+    }
+
+    #[test]
+    fn test_sequence_cubic_i32()
+    {
+        let expected = vec![1, 8, 27, 64, 125];
+        let n = expected.len();
+        let actual = Sequence::Cubic.vec(n);
+        assert_eq!(expected, actual);
+    }
+
+    #[test]
     fn test_sequence_fibonacci_i32()
     {
         let expected = vec!
@@ -309,6 +408,27 @@ mod tests
         ];
         let n = expected.len();
         let actual = Sequence::LookAndSay(1).vec(n);
+        assert_eq!(expected, actual);
+    }
+
+    #[test]
+    fn test_sequence_map_i32()
+    {
+        // Natural * 2 = Even
+        let expected = Sequence::<i32>::Even(true).vec(4);
+        let actual = Sequence::<i32>::Natural(true).map(4, |x| x * 2);
+        assert_eq!(expected, actual);
+    }
+
+    #[test]
+    fn test_sequence_fold_i32()
+    {
+        // Triangular.Fold(2, sum) = Square
+        let expected = Sequence::<i32>::Square.vec(4);
+        let actual = Sequence::<i32>::Triangular.fold(4, -1, 1, |v|
+        {
+            v.iter().map(|x| x.unwrap_or(0)).sum()
+        });
         assert_eq!(expected, actual);
     }
 }